fix(ui): added no permission placeholder instead 404
