+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/zh-cn.json
@@ -1206,6 +1206,7 @@
     "queries-result-test": "Queries returning 1 or more rows will result in the test failing.",
     "query-log-duration-message": "Configuration to tune how far we want to look back in query logs to process usage data.",
     "query-used-by-other-tables": "Query used by other tables",
+    "reach-out-to-admin-for-access": "Please reach out to your Admin to get access.",
     "reacted-with-emoji": "reacted with {{type}} emoji",
     "redirect-message": "Please wait while you are being redirected.",
     "redirecting-to-home-page": "Redirecting to the home page",
