+++ openmetadata-ui/src/main/resources/ui/src/pages/GlobalSettingPage/GlobalSettingPage.tsx
@@ -11,14 +11,111 @@
  *  limitations under the License.
  */
 
+import { Menu, MenuProps } from 'antd';
+import { ItemType } from 'antd/lib/menu/hooks/useItems';
+import LeftPanelCard from 'components/common/LeftPanelCard/LeftPanelCard';
 import PageContainerV1 from 'components/containers/PageContainerV1';
-import GlobalSetting from 'components/GlobalSetting/GlobalSetting';
-import React from 'react';
+import PageLayoutV1 from 'components/containers/PageLayoutV1';
+import { usePermissionProvider } from 'components/PermissionProvider/PermissionProvider';
+import GlobalSettingRouter from 'components/router/GlobalSettingRouter';
+import { GlobalSettingOptions } from 'constants/GlobalSettings.constants';
+import { ELASTIC_SEARCH_RE_INDEX_PAGE_TABS } from 'enums/ElasticSearch.enum';
+import { TeamType } from 'generated/entity/teams/team';
+import { useAuth } from 'hooks/authHooks';
+import React, { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useHistory, useParams } from 'react-router-dom';
+import {
+  getGlobalSettingMenuItem,
+  getGlobalSettingsMenuWithPermission,
+  MenuList,
+} from 'utils/GlobalSettingsUtils';
+import {
+  getSettingPath,
+  getSettingsPathWithFqn,
+  getTeamsWithFqnPath,
+} from 'utils/RouterUtils';
+import './global-setting-page.style.less';
 
 const GlobalSettingPage = () => {
+  const history = useHistory();
+  const { t } = useTranslation();
+  const { tab, settingCategory } = useParams<{ [key: string]: string }>();
+
+  const { permissions } = usePermissionProvider();
+
+  const { isAdminUser } = useAuth();
+
+  const menuItems: ItemType[] = useMemo(
+    () =>
+      getGlobalSettingsMenuWithPermission(permissions, isAdminUser).reduce(
+        (acc: ItemType[], curr: MenuList) => {
+          const menuItem = getGlobalSettingMenuItem({
+            label: curr.category,
+            key: curr.key,
+            category: curr.category,
+            children: curr.items,
+            type: 'group',
+            isBeta: curr.isBeta,
+          });
+          if (menuItem.children?.length) {
+            return [...acc, menuItem];
+          } else {
+            return acc;
+          }
+        },
+        [] as ItemType[]
+      ),
+    [permissions]
+  );
+
+  const onClick: MenuProps['onClick'] = (e) => {
+    // As we are setting key as "category.option" and extracting here category and option
+    const [category, option] = e.key.split('.');
+
+    switch (option) {
+      case GlobalSettingOptions.TEAMS:
+        history.push(getTeamsWithFqnPath(TeamType.Organization));
+
+        break;
+      case GlobalSettingOptions.SEARCH:
+        history.push(
+          getSettingsPathWithFqn(
+            category,
+            option,
+            ELASTIC_SEARCH_RE_INDEX_PAGE_TABS.ON_DEMAND
+          )
+        );
+
+        break;
+      default:
+        history.push(getSettingPath(category, option));
+
+        break;
+    }
+  };
+
+  const leftPanel = menuItems.length ? (
+    <LeftPanelCard id="settings">
+      <Menu
+        className="custom-menu"
+        data-testid="global-setting-left-panel"
+        items={menuItems}
+        mode="inline"
+        selectedKeys={[`${settingCategory}.${tab}`]}
+        onClick={onClick}
+      />
+    </LeftPanelCard>
+  ) : null;
+
   return (
     <PageContainerV1>
-      <GlobalSetting />
+      <PageLayoutV1
+        className="tw-h-full tw-px-6"
+        leftPanel={leftPanel}
+        pageTitle={t('label.setting-plural')}>
+        <GlobalSettingRouter />
+      </PageLayoutV1>
     </PageContainerV1>
   );
 };
