+++ openmetadata-ui/src/main/resources/ui/src/pages/GlobalSettingPage/global-setting-page.style.less
