+++ openmetadata-ui/src/main/resources/ui/src/pages/services/ServicesPage.tsx
@@ -23,7 +23,10 @@ import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import { getServices } from 'rest/serviceAPI';
 import { pagingObject, SERVICE_VIEW_CAP } from '../../constants/constants';
-import { NO_PERMISSION_TO_VIEW } from '../../constants/HelperTextUtil';
+import {
+  NO_PERMISSION_TO_VIEW,
+  REACH_OUT_TO_ADMIN_FOR_ACCESS,
+} from '../../constants/HelperTextUtil';
 import {
   OPEN_METADATA,
   SERVICE_CATEGORY,
@@ -119,7 +122,9 @@ const ServicesPage = () => {
     <Row>
       <Col span={24}>
         <ErrorPlaceHolder>
-          <p>{NO_PERMISSION_TO_VIEW}</p>
+          <p className="text-center">
+            {NO_PERMISSION_TO_VIEW} <br /> {REACH_OUT_TO_ADMIN_FOR_ACCESS}
+          </p>
         </ErrorPlaceHolder>
       </Col>
     </Row>
